Potentially workable solution
This is an untested but potentially complete version. It no longer creates a profileData.txt file but puts the data directly into a format suitable for replacing the existing data in the json file. It also empties the temp folder before starting (but doesn't allow the user the option to bail out if it's not already empty)