Very alpha!
We are now creating a usable sb3 file... almost/ At the moment the imported png and svg files are appearing at what appears to be half size. Also of note is that we are now creating a unique temp folder to do all the work in and that folder is automatically removed afterwards.