Working version!
Using regex rather than re so that I can search for matching square brackets. Modified various parameters to get the importing of the png and svg files correct for size and positioning. This now achieves all "basic" goals in so much as everything that was previously needed to be done is all automated.